rebased with cpdnpq-1338-refactor-forms-folder
