+++ app/services/ecf/ecf_application_synchronization.rb
@@ -1,24 +1,47 @@
 module Ecf
   class EcfApplicationSynchronization
+    PER_PAGE = 50
+    DEFAULT_LIMIT = 1000
+
+    def initialize(limit = DEFAULT_LIMIT)
+      @limit = limit
+    end
+
     def call
-      uri = build_uri
-      request = build_http_get_request(uri)
-      response = send_http_request(uri, request)
-      handle_response(response)
+      process_batch_applications
     rescue StandardError => e
       Rails.logger.error "An error occurred during application synchronization: #{e.message}"
     end
 
   private
 
+    # Remove this code once all application statuses have been migrated.
+    def process_batch_applications
+      applications_to_sync.find_in_batches(batch_size: PER_PAGE) do |group|
+        batch_runner(group.pluck(:ecf_id))
+      end
+    end
+
+    def batch_runner(batch)
+      uri = build_uri
+      request = build_http_get_request(uri, batch)
+      response = send_http_request(uri, request)
+      handle_response(response)
+    end
+
     def build_uri
       URI.parse("#{ENV['ECF_APP_BASE_URL']}/api/v1/npq/application_synchronizations")
     end
 
-    def build_http_get_request(uri)
+    def build_http_get_request(uri, ecf_ids)
+      uri.query = "ecf_ids=#{ecf_ids.join(',')}"
       request = Net::HTTP::Get.new(uri)
       request["Authorization"] = "Bearer #{ENV['ECF_APP_BEARER_TOKEN']}"
       request
+    end
+
+    def applications_to_sync
+      Application.where(lead_provider_approval_status: nil).limit(@limit)
     end
 
     def send_http_request(uri, request)
