+++ app/services/feature.rb
@@ -2,6 +2,7 @@ class Feature
   REGISTRATION_OPEN_DATE = Time.zone.parse("6 June 2022 12:00")
 
   REGISTRATION_CLOSED_KEY = "Registration closed".freeze
+  REGISTRATION_DISABLED = "Registration disabled".freeze
 
   FEATURE_FLAG_KEYS = [
     REGISTRATION_CLOSED_KEY,
@@ -28,6 +29,22 @@ class Feature
 
     def registration_closed?
       Flipper.enabled?(REGISTRATION_CLOSED_KEY)
+    end
+
+    def registration_disabled?
+      Flipper.enabled?(REGISTRATION_DISABLED)
+    end
+
+    def registration_enabled?
+      !Flipper.enabled?(REGISTRATION_DISABLED)
+    end
+
+    def disable_registration!
+      Flipper.enable(REGISTRATION_DISABLED)
+    end
+
+    def enable_registration!
+      Flipper.disable(REGISTRATION_DISABLED)
     end
   end
 end
